download cover

+++ utility.py
@@ -8,6 +8,8 @@ def download_image(img_src_url: str, img_relative_path=None) -> str:
     if not img_relative_path:
         img_relative_path = f"images/{img_src_url.split('/')[-3]}/{img_src_url.split('/')[-2]}/{img_src_url.split('/')[-1]}"
     os.makedirs(os.path.dirname(img_relative_path), exist_ok=True)
+    if os.path.exists(img_relative_path):
+        return
     with open(img_relative_path, "wb") as f:
         f.write(response.content)
